fix; adapted render command; add intermediates_dir
Now the command to compile the input markdown includes some extra args.
Most important now the arg 'intermediates_dir' is included and set to
OUTDIR.
Also argument 'clean' is set to FALSE, to retain the intermediate
markdown file.
'run_pandoc' is set to TRUE, but this is also the default.